kids_learning_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
google-cloud-storage
